Introduced new test_beta and study_beta functions.
These will operate on continous (double typed) memory and threshold matrices, and updat
them with draws from the beta distribution. The beta is parameterized with
its two shape parameters, but are constrained to have same mean and sd as a binomial distrution, where the
binomial is parameterized by the number of features/threshold and relevant model parameters like LR, TR and FR.

study() and test() are now wrappers for the function you really want (*_binomial or *beta) and will
call either one based on the type of data in the matrix. These should be considered a deprecated compatability
layer to avoid breaking old models.