now use dplyr::first to help build the vector for correct-only RT's in PCL::freeRecall().
This has the advantage of specifiying a default value (set here to numeric(0)) when there is not first item! This at first seems like an unlikely scenario to cover for, but not when you consider that the probability that at least 1 of the 1000 simulations will have all features below threshold.