Smarter closures for PR/CR functions.
Instead of having a factory for each one, use one factory for all PR/CR function that returns a list of functions based on the distribution type.

When updating the PCR parameters, get the enclosing environment and modify it to have the new parameters directly instead of re-running the factory to enclose them.