Changed how PCR tracks items practice histories.
Practices history is now stored in a data frame, with a column identifying the cue, and a column identifying the practice method. Each row is a practice opportunity.

Since I had changed the summary method to deal with this, I also made it deal with objects that never had a test trial more efficiently.